mcp_server/requirements.txt
mcp
aiohttp
soundfile
numpy
scipy
pyloudnorm
pytest-asyncio
